refactor(api): move expect-job helpers into chai-custom-helpers folder
